pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestry"
version = "1.0.0"
description = "Orchestry - Automatic orchestration, scaling and load balancing for containerized applications"
authors = [{name = "Hemanth S Banur", email = "arjunbanur27@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer==0.17.4",
    "requests==2.32.5",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "platformdirs==4.4.0"
]

[tool.setuptools]
packages = [
    "cli"
]

[project.scripts]
orchestry = "cli.main:app"

[project.urls]
Homepage = "https://arjuuuuunnnnn.github.io/Orchestry/"
Repository = "https://github.com/arjuuuuunnnnn/Orchestry"
